+++ Content.Client/_NF/Radar/RadarBlipSystem.cs
@@ -1,19 +1,27 @@
 using System.Numerics;
-using Content.Shared._Mono.Radar;
+using Content.Shared._NF.Radar;
 using Robust.Shared.Timing;
 
-namespace Content.Client._Mono.Radar;
+namespace Content.Client._NF.Radar;
 
-public sealed partial class RadarBlipsSystem : EntitySystem
+/// <summary>
+/// A system for requesting, receiving, and caching radar blips.
+/// Sends off ad hoc requests for blips, caches them for a period of time, and draws them when requested.
+/// </summary>
+/// <remarks>
+/// Ported from Monolith's RadarBlipsSystem.
+/// </remarks>
+public sealed partial class RadarBlipSystem : EntitySystem
 {
     private const double BlipStaleSeconds = 3.0;
     private static readonly List<(Vector2, float, Color, RadarBlipShape)> EmptyBlipList = new();
     private static readonly List<(NetEntity? Grid, Vector2 Position, float Scale, Color Color, RadarBlipShape Shape)> EmptyRawBlipList = new();
     private TimeSpan _lastRequestTime = TimeSpan.Zero;
-    private static readonly TimeSpan RequestThrottle = TimeSpan.FromMilliseconds(250);
+    // Minimum time between requests.  Slightly larger than the server-side value.
+    private static readonly TimeSpan RequestThrottle = TimeSpan.FromMilliseconds(1250);
 
     // Maximum distance for blips to be considered visible
-    private const float MaxBlipRenderDistance = 1000f;
+    private const float MaxBlipRenderDistance = 256f;
 
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly SharedTransformSystem _xform = default!;
@@ -28,6 +36,9 @@ public sealed partial class RadarBlipsSystem : EntitySystem
         SubscribeNetworkEvent<GiveBlipsEvent>(HandleReceiveBlips);
     }
 
+    /// <summary>
+    /// Handles receiving blip data from the server.
+    /// </summary>
     private void HandleReceiveBlips(GiveBlipsEvent ev, EntitySessionEventArgs args)
     {
         if (ev?.Blips == null)
@@ -39,23 +50,21 @@ public sealed partial class RadarBlipsSystem : EntitySystem
         _lastUpdatedTime = _timing.CurTime;
     }
 
+    /// <summary>
+    /// Requests blip data from the server for the given radar console, throttled to avoid spamming.
+    /// </summary>
     public void RequestBlips(EntityUid console)
     {
-        // Only request if we have a valid console
         if (!Exists(console))
             return;
 
-        // Add request throttling to avoid network spam
         if (_timing.CurTime - _lastRequestTime < RequestThrottle)
             return;
 
         _lastRequestTime = _timing.CurTime;
 
         // Cache the radar position for distance culling
-        if (TryComp<TransformComponent>(console, out var xform))
-        {
-            _radarWorldPosition = _xform.GetWorldPosition(console);
-        }
+        _radarWorldPosition = _xform.GetWorldPosition(console);
 
         var netConsole = GetNetEntity(console);
         var ev = new RequestBlipsEvent(netConsole);
@@ -68,18 +77,14 @@ public sealed partial class RadarBlipsSystem : EntitySystem
     /// </summary>
     public List<(Vector2, float, Color, RadarBlipShape)> GetCurrentBlips()
     {
-        // If it's been more than the stale threshold since our last update,
-        // the data is considered stale - return an empty list
         if (_timing.CurTime.TotalSeconds - _lastUpdatedTime.TotalSeconds > BlipStaleSeconds)
             return EmptyBlipList;
 
         var result = new List<(Vector2, float, Color, RadarBlipShape)>(_blips.Count);
-
         foreach (var blip in _blips)
         {
             Vector2 worldPosition;
 
-            // If no grid, position is already in world coordinates
             if (blip.Grid == null)
             {
                 worldPosition = blip.Position;
@@ -91,15 +96,10 @@ public sealed partial class RadarBlipsSystem : EntitySystem
                 result.Add((worldPosition, blip.Scale, blip.Color, blip.Shape));
                 continue;
             }
-
-            // If grid exists, transform from grid-local to world coordinates
             if (TryGetEntity(blip.Grid, out var gridEntity))
             {
-                // Transform the grid-local position to world position
                 var worldPos = _xform.GetWorldPosition(gridEntity.Value);
                 var gridRot = _xform.GetWorldRotation(gridEntity.Value);
-
-                // Rotate the local position by grid rotation and add grid position
                 var rotatedLocalPos = gridRot.RotateVec(blip.Position);
                 worldPosition = worldPos + rotatedLocalPos;
 
@@ -110,7 +110,6 @@ public sealed partial class RadarBlipsSystem : EntitySystem
                 result.Add((worldPosition, blip.Scale, blip.Color, blip.Shape));
             }
         }
-
         return result;
     }
