+++ Content.Server/_NF/Radar/RadarBlipComponent.cs
@@ -1,9 +1,9 @@
-namespace Content.Server._Mono.Radar;
+using Content.Shared._NF.Radar;
 
-using Content.Shared._Mono.Radar;
+namespace Content.Server._NF.Radar;
 
 /// <summary>
-/// These handle objects which should be represented by radar blips.
+/// Handles objects which should be represented by radar blips.
 /// </summary>
 [RegisterComponent]
 public sealed partial class RadarBlipComponent : Component
@@ -11,39 +11,42 @@ public sealed partial class RadarBlipComponent : Component
     /// <summary>
     /// Color that gets shown on the radar screen.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField("radarColor")]
-    public Color RadarColor = Color.Red;
+    [DataField]
+    public Color RadarColor { get; set; } = Color.Red;
 
     /// <summary>
     /// Color that gets shown on the radar screen when the blip is highlighted.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField("highlightedRadarColor")]
-    public Color HighlightedRadarColor = Color.OrangeRed;
+    [DataField]
+    public Color HighlightedRadarColor { get; set; } = Color.OrangeRed;
 
     /// <summary>
     /// Scale of the blip.
     /// </summary>
     [DataField]
-    public float Scale = 1;
+    public float Scale { get; set; } = 1f;
 
     /// <summary>
     /// The shape of the blip on the radar.
     /// </summary>
     [DataField]
-    public RadarBlipShape Shape = RadarBlipShape.Circle;
+    public RadarBlipShape Shape { get; set; } = RadarBlipShape.Circle;
 
     /// <summary>
     /// Whether this blip should be shown even when parented to a grid.
     /// </summary>
     [DataField]
-    public bool RequireNoGrid = false;
+    public bool RequireNoGrid { get; set; } = false;
 
     /// <summary>
     /// Whether this blip should be visible on radar across different grids.
     /// </summary>
     [DataField]
-    public bool VisibleFromOtherGrids = false;
+    public bool VisibleFromOtherGrids { get; set; } = true;
 
+    /// <summary>
+    /// Whether this blip is enabled and should be shown on radar.
+    /// </summary>
     [DataField]
-    public bool Enabled = true;
+    public bool Enabled { get; set; } = true;
 }
