--- Resources/Prototypes/Entities/Objects/Vehicles/actions.yml
@@ -1,10 +1,0 @@
-﻿- type: entity
-  id: ActionVehicleHorn
-  name: Honk
-  description: Honk!
-  categories: [ HideSpawnMenu ]
-  components:
-  - type: InstantAction
-    useDelay: 3.4
-    icon: Objects/Fun/bikehorn.rsi/icon.png
-    event: !type:HonkActionEvent
