--- Resources/Prototypes/Entities/Objects/Vehicles/keys.yml
@@ -1,86 +1,0 @@
-- type: entity
-  parent: BaseItem
-  id: VehicleKeyJanicart
-  name: janicart keys
-  description: Interesting design.
-  components:
-  - type: Item
-    size: Tiny
-  - type: Tag
-    tags:
-    - VehicleKey
-    - JanicartKeys
-  - type: Sprite
-    sprite: Objects/Vehicles/janicart.rsi
-    state: keys
-
-- type: entity
-  parent: VehicleKeyJanicart
-  id: VehicleKeySecway
-  name: secway keys
-  description: The keys to the future.
-  components:
-  - type: Tag
-    tags:
-    - VehicleKey
-    - SecwayKeys
-  - type: Sprite
-    sprite: Objects/Vehicles/secway.rsi
-    state: keys
-
-- type: entity
-  parent: VehicleKeyJanicart
-  id: VehicleKeyATV
-  name: ATV keys
-  description: Think this looks like just one key? ATV keys means "actually two vehicle keys."
-  components:
-  - type: Tag
-    tags:
-    - VehicleKey
-    - ATVKeys
-  - type: Sprite
-    sprite: Objects/Vehicles/atv.rsi
-    state: keys
-
-- type: entity
-  parent: VehicleKeyATV
-  id: VehicleKeySkeleton
-  name: vehicle skeleton keys
-  description: Unlock any vehicle.
-  components:
-  - type: Tag
-    tags:
-    - VehicleKey
-    - JanicartKeys
-    - SecwayKeys
-    - ATVKeys
-    - SyndicateSegwayKeys
-    - SkeletonMotorcycleKeys
-
-- type: entity
-  parent: VehicleKeyJanicart
-  id: VehicleKeySyndicateSegway
-  name: syndicate segway keys
-  description: Patterned after the iconic EMAG design.
-  components:
-    - type: Tag
-      tags:
-        - VehicleKey
-        - SyndicateSegwayKeys
-    - type: Sprite
-      sprite: Objects/Vehicles/syndicatesegway.rsi
-      state: keys
-
-- type: entity
-  parent: VehicleKeyATV
-  id: VehicleKeySkeletonMotorcycle
-  name: skeleton motorcycle keys
-  description: A beautiful set of keys adorned with a skull.
-  components:
-    - type: Tag
-      tags:
-        - VehicleKey
-        - SkeletonMotorcycleKeys
-    - type: Sprite
-      sprite: Objects/Vehicles/motorbike.rsi
-      state: keys
